receiver/requirements.txt
fastapi==0.68.2
fastapi-utils==0.2.1
pydantic==1.10.2
pydantic[dotenv]
uvicorn
requests
pycryptodome
pandas
openpyxl
fastecdsa
pwntools
pyelftools==0.29